- SwaggerConfig add DI - ConfigureSwaggerOptions moved to a Configurations folder - Removed unused using - ConfigureSwaggerOptions refactoring
